pyproject.toml
[project]
name = "sample-agentframework-agent"
version = "0.1.0"
description = "Sample Agent Framework Agent using Microsoft Agent 365 SDK"
authors = [
    { name = "Microsoft", email = "support@microsoft.com" }
]
dependencies = [
    # AgentFramework SDK - The official package
    "agent-framework-azure-ai",
    
    # Azure AI Projects - explicitly require pre-release version
    "azure-ai-agents>=1.2.0b5",
    "azure-ai-projects>=2.0.0b2",

    # Microsoft Agents SDK - Official packages for hosting and integration
    "microsoft-agents-hosting-aiohttp",
    "microsoft-agents-hosting-core",
    "microsoft-agents-authentication-msal",
    "microsoft-agents-activity",

    # Azure SDK components
    "azure-identity",

    # Core dependencies
    "python-dotenv",
    "aiohttp",

    # HTTP server support for MCP servers
    "uvicorn[standard]>=0.20.0",
    "fastapi>=0.100.0",

    # HTTP client
    "httpx>=0.24.0",

    # Data validation
    "pydantic>=2.0.0",

    # Additional utilities
    "typing-extensions>=4.0.0",

    # Microsoft Agent 365 SDK packages (pre-release versions — this is a POC)
    "microsoft_agents_a365_tooling >= 0.1.0.dev0",
    "microsoft_agents_a365_tooling_extensions_agentframework >= 0.1.0.dev0",
    "microsoft_agents_a365_observability_core >= 0.2.0.dev0",
    "microsoft_agents_a365_observability_extensions_agent_framework >= 0.1.0.dev0",
    "microsoft_agents_a365_runtime >= 0.1.0.dev0",
    "microsoft_agents_a365_notifications >= 0.1.0.dev0",

    # Microsoft OpenTelemetry Distro (vendored wheel — installed with --no-deps to skip semantic-kernel)
    # Run: pip install --no-deps vendor/microsoft_opentelemetry-0.1.0b1-py3-none-any.whl
    # "microsoft-opentelemetry @ file:vendor/microsoft_opentelemetry-0.1.0b1-py3-none-any.whl",

    # Distro dependencies we actually use (manually listed since we skip the wheel's own deps)
    "opentelemetry-sdk~=1.39",
    "azure-monitor-opentelemetry~=1.6.0",
    "opentelemetry-exporter-otlp-proto-http~=1.39",
    "opentelemetry-instrumentation-django~=0.60b0",
    "opentelemetry-instrumentation-fastapi~=0.60b0",
    "opentelemetry-instrumentation-flask~=0.60b0",
    "opentelemetry-instrumentation-psycopg2~=0.60b0",
    "opentelemetry-instrumentation-requests~=0.60b0",
    "opentelemetry-instrumentation-urllib~=0.60b0",
    "opentelemetry-instrumentation-urllib3~=0.60b0",
    "opentelemetry-instrumentation-openai-v2~=2.3b0",
    "opentelemetry-instrumentation-openai-agents~=0.53.0",
    "opentelemetry-instrumentation-langchain~=0.53.0",
    "opentelemetry-resource-detector-azure>=0.1.5,<1.0.0",
    "microsoft-agents-a365-observability-extensions-openai>=0.1.0.dev0",
    "microsoft-agents-a365-observability-extensions-langchain>=0.1.0.dev0",
    "microsoft-agents-a365-observability-extensions-agent-framework>=0.1.0.dev0",
]
requires-python = ">=3.11"

# Package index configuration
# PyPI is the default/primary source, local packages are fallback
[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

# Allow pre-release versions for Microsoft Agent 365 SDK packages
# This ensures we always get the latest features and fixes
[tool.uv]
prerelease = "allow"
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
